Fix adding new item

+++ src/Shared/ItemInfo.cs
@@ -6,12 +6,12 @@ namespace Marketplace.Shared
 {
     public class ItemInfo
     {
-        public ushort ItemId { get; set; }
+        public int ItemId { get; set; }
         public byte Quality { get; set; }
         public byte Amount { get; set; }
         public byte[] Metadata { get; set; }
 
-        public static ItemInfo Create(ushort id, byte quality, byte amount, byte[] metadata)
+        public static ItemInfo Create(int id, byte quality, byte amount, byte[] metadata)
         {
             return new ItemInfo()
             {
